Dataset now stores all genotypical information

+++ src/epistasis/dataset.h
@@ -3,7 +3,6 @@
 
 #include <stdint.h>
 #include <stdlib.h>
-// #include <sys/types.h>
 
 #include <bioformats/vcf/vcf_file_structure.h>
 #include <containers/array_list.h>
@@ -19,14 +18,14 @@ epistasis_dataset *epistasis_dataset_new();
 void epistasis_dataset_free(epistasis_dataset *dataset);
 
 
-void epistasis_dataset_process_records(vcf_record_t **variants, int num_variants, int num_samples, int *phenotypes, epistasis_dataset *dataset);
+void epistasis_dataset_process_records(vcf_record_t **variants, size_t num_variants, int num_samples, int *phenotypes, epistasis_dataset *dataset);
 
 void epistasis_dataset_add_entry(uint16_t *genotype_count, epistasis_dataset *dataset);
 
 
-uint16_t epistasis_dataset_get_num_variants(epistasis_dataset *dataset);
+size_t epistasis_dataset_get_num_variants(epistasis_dataset *dataset);
 
-uint16_t *epistasis_dataset_get_variant_counts(size_t index, epistasis_dataset *dataset);
+uint8_t *epistasis_dataset_get_variant_counts(size_t index, epistasis_dataset *dataset);
 
 
 #endif
